BinaryTree
In this project we were working with visitor pattern which allow user to  add new method into binary tree program without modifying the original the code. The two new method that was added was to find the max of the binary tree and print the binary tree in post traversal order.